Fix coachSection issue
Apparently the image was flipped. I basically dragged one of the
corners while editing the size of the Rect in a way that the corner
I was dragging overpassed another corner, and the image flipped
back to normal orientation.
The positions was also outside the frame, So I moved it.