Add highlight to player's farm
I added a png file based on the normal farm and edited to have a yellow
highlight. I decided to change this tile in the GameControllerScript in
the same way that pest is being updated. I never used the tilemap
from unity before and I was struggling to edit in the Unity editor.
Also, in case we want to change the location of the player's farm, I
believe it is easier to edit from the script.