Change tutorial steps
I fix the content of the tutorial based on Jennifer's email with
information about the game. I adjusted the position of the popups and of
the arrow highlighting  the GUI. Design and style can be improved of course.
My main concern was to make it scalable to have more or less tutorial
steps if needed.