Fix bug on state machine
Because I added the Init state plus the Tutorial state, after reaching
PreparingNextYear for the first time, the NextState method was going to
Init state again. I added a new method to start a new round that sets
the state to WaitingForPlayerInput. After preparing a NewYear, it is
either calling EndGame or StartRound. If you prefer to keep the state
machine clean with only states from the game, I can turn the tutorial
phase into a boolean or something else outside the state machine.